Switch to semantic versioning (#117)
Instead of an incrementing build number we use [semantic
versioning](http://semver.org) from now on. Starting at ``0.10.0``.
As far as I am able to figure from the non existing documentation of
``metadata.json`` keys there are no particular contrains on the version
string, so this switch should be trivial.

Closes: #117