Provide basic documentation infrastructure(#119)
This commit provides a basic sphinx based infrastructure to handle our
documentation needs. Besides setting up sphinx itself we did the
following:

- Create several new ``make`` targets for easy building, opening and
  testing of the documentation. As we depend on some additional python
  packages for this, we added a ``requirements.pip`` file used by the
  new ``make develop`` target (which is responsible for setting up the
  entire toolchain.
- ``README.md`` has been converted to ``rst`` in order to be in line
  with the rest of the project. We also updated and reordered some of
  its content. For that purpose versioning and release related
   information has been moved to dedicated stub pages.
- A ``setup.cfg`` has been added to store additional settings for some
  of our tools.

Testing of the documentation is done by building it in a dedicated
``_test_build`` directory for now. Once a clear testing/CI process
emerges this may be optimised.

We used the opportunity to add ``bumpversion`` to the toolchain while we
were at it.

Closes: #119