Fix automatic refresh of the panel widget
This widget installs a GSourceFunc to be triggered on a 60 seconds timeout
but it does not explicitly return GLib.SOURCE_CONTINUE in the callback,
meaning that the source will be removed after it gets called once, leaving
the label outdated forever.

This fixes this problem by explicitly returning GLib.SOURCE_CONTINUE there.

https://github.com/projecthamster/hamster-shell-extension/issues/189