Define exportable properties with 'var'
This is to fix this new gjs warning:

```
Some code accessed the property 'PanelWidget' on the module
'panelWidget'. That property was defined with 'let' or 'const' inside
the module. This was previously supported, but is not correct according
to the ES6 standard. Any symbols to be exported from a module must be
defined with 'var'. The property access will work as previously for the
time being, but please fix your code anyway.
```

Ref.: #198 #205