Avoid crashes due to accessing destroyed objects
The crashes that I have been experiencing all have backtraces leading
back to the "refresh" event and error messages explaining that we are
trying to use objects that have been destroyed.

This is the only place where we are destroying objects explicitly
instead of relying on the garbage collector. To fix the crash, I'm
simply using the `remove_all_children()` method instead of the
`destroy_all_children()` one.

Hopefully fixes #212