metadata.json: remove "version" field
The "version" field in the extension meta data is reserved for internal
use on extensions.gnome.org. If at all, the version should be set to a
value that can be used as an argument to python's int() function, which
is not the case for the string we currently use.
The general recommendation is not to use "version" in metadata.json.

See e.g. https://gitlab.gnome.org/Infrastructure/extensions-web/issues/102#note_740134