Makefile: configurable extension UUID
We seem to have issues with the contact@projecthamster.org UUID.
GNOME shell 3.36 looks it up on extensions.gnome.org, finds
it to be "blacklisted", disables and deletes (!) it.
move data/metadata.json to data/metadata.json.in

So, make it possible to choose another UUID during build.
Simply run:

     make UUID=my-nice-uuid@some.domain

This requires to create metadata.json from metadata.json.in.