Stop tracking when the screensaver cuts in.
Listen to the ScreenSaver dbus signal to know when to stop the current activity.

Ideally it would backout the idle time before the screensaver cuts in, but I can't see any way to know if it cut in by idle activity or a key shortcut (in which case there should be no time backed out).