panelWidget: Remove show() method
This widget defined a `show()` method, which would show the menu.
However, `show()` is already a method defined by some class higher up in
the hierarchy, which makes the *widget* in the status area visible.  By
defining a `show()` method here, the behavior of the `show()` method
changes, causing issues.

One such issue showed when combining this extension with the
status-area-horizontal-spacing extension (prior to v2.6), which calls
`hide()` and `show()` on each widget in the status area. With the
overridden `show()` method in the hamster extension, this would hide the
widget and then show the menu, breaking the extension because the widget
would stay hidden.

Since the `show()` method was not actuall used anywhere, this commit
removes it to prevent the conflict.