Report errors on initial DBUS connection
When either dbus proxy cannot be created (i.e. hamster is not running
and cannot be autostarted), report that to the user.

This only reports these errors, not other handling is changed (on error,
the proxy objects passed will be undefined or null, so further
initialization of the extension should already be skipped). Making error
handling more robust (and e.g. recovering of hamster is started later)
would be nice, but is too complex to handle now.