ongoingFactEntry: Remove custom style class
This fixes the "Enter activity.." text input to actually be usable and visible
on my Ubuntu 24.04 (GNOME 46) when using light mode.

In dark mode having the extra style class seems to work fine, but in light mode
the input field is a light color and the text is also almost in the same color,
so you cannot see what you write.